requirements-student.txt
torch>=2.4.0
torchvision>=0.19.0
transformers>=4.55.0
peft>=0.13.0
accelerate>=0.34.0
huggingface-hub>=0.25.0
safetensors>=0.4.5
Pillow>=10.4.0
tqdm>=4.66.0
